requirements.txt
# Core dependencies
aiohttp>=3.8.0
pyyaml>=6.0

# Language detection
langdetect>=1.0.9

# System notifications (Windows)
win10toast>=0.9; sys_platform == 'win32'

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Optional: Cloud translation
# requests>=2.28.0

# Optional: Local AI
# ollama>=0.1.0
